add custom wait time for retries
